First PoC with Layer and Source classes
