+++ cartoframes/carto_vl/source/source.py
@@ -1,0 +1,9 @@
+from __future__ import absolute_import
+
+class Source(object):
+    """
+      TODO
+    """
+
+    def __init__(self, query):
+        self.query = query
